+++ core/src/main/java/org/elkd/core/runtime/client/consumer/ConsumerACK.kt
@@ -1,0 +1,16 @@
+package org.elkd.core.runtime.client.consumer
+
+import org.elkd.core.server.client.RpcConsumerResponse
+
+object ConsumerACK {
+  enum class Codes(val id: Int) {
+    OK(0),
+    OPERATION_INVALID(1),
+    UNKNOWN_TOPIC(2)
+  }
+
+  object Rpcs {
+    val UNKNOWN_TOPIC: RpcConsumerResponse = RpcConsumerResponse.newBuilder().setStatus(Codes.UNKNOWN_TOPIC.id).build()
+    val OPERATION_INVALID: RpcConsumerResponse = RpcConsumerResponse.newBuilder().setStatus(Codes.OPERATION_INVALID.id).build()
+  }
+}