+++ core/src/main/java/org/elkd/core/runtime/client/consumer/ConsumerRequest.kt
@@ -1,0 +1,8 @@
+package org.elkd.core.runtime.client.consumer
+
+import org.elkd.core.runtime.topic.Topic
+
+data class ConsumerRequest(
+    val topic: Topic,
+    val index: Long?
+)