+++ core/src/main/java/org/elkd/core/runtime/client/stream/StreamObserverDecorator.kt
@@ -3,20 +3,20 @@ package org.elkd.core.runtime.client.stream
 import io.grpc.stub.StreamObserver
 
 class StreamObserverDecorator<T>(
-    val stream: StreamObserver<T>,
+    val next: StreamObserver<T>,
     val onNextBlock: (StreamObserver<T>, T) -> Unit = { stream, value -> stream.onNext(value) },
     val onErrorBlock: (StreamObserver<T>, Throwable) -> Unit = { stream, throwable -> stream.onError(throwable) },
-    val onCompleteBlock: (StreamObserver<T>) -> Unit = { stream.onCompleted() }
+    val onCompleteBlock: (StreamObserver<T>) -> Unit = { next.onCompleted() }
 ) : StreamObserver<T> {
   override fun onNext(value: T) {
-    onNextBlock(stream, value)
+    onNextBlock(next, value)
   }
 
   override fun onError(t: Throwable) {
-    onErrorBlock(stream, t)
+    onErrorBlock(next, t)
   }
 
   override fun onCompleted() {
-    onCompleteBlock(stream)
+    onCompleteBlock(next)
   }
 }